requirements-dev.txt
# Development and testing dependencies
# Install with: pip install -r requirements-dev.txt

# Include base requirements
-r requirements.txt

# Code quality
black>=23.12.0
flake8>=7.0.0
isort>=5.13.0

# Testing
pytest-xdist>=3.5.0  # Run tests in parallel
factory-boy>=3.3.0   # Test data factories
faker>=22.0.0        # Generate fake data for tests

# Debugging
ipdb>=0.13.13
django-debug-toolbar>=4.2.0

# Documentation
sphinx>=7.2.6
sphinx-rtd-theme>=2.0.0
